Have fleshed out EditSubscriptionActivity more; Allowed functionality for getting data off of intents and added in comments.
